Refactor llm_chat_task
