+++ src/zrb/llm/task/chat/task.py
@@ -387,30 +387,6 @@ class LLMChatTask(BuilderMixin, RunnerMixin, BaseTask):  # type: ignore[reportIn
         self._show_ollama_models = show_ollama_models
         self._show_pydantic_ai_models = show_pydantic_ai_models
 
-    @property
-    def llm_config(self) -> LLMConfig:
-        return self._llm_config
-
-    @property
-    def llm_limiter(self) -> LLMLimiter | None:
-        return self._llm_limiter
-
-    @property
-    def permissions(self) -> "PermissionPolicyInput":
-        return self._permissions
-
-    @permissions.setter
-    def permissions(self, value: "PermissionPolicyInput"):
-        self._permissions = value
-
-    @property
-    def sandbox(self) -> "SandboxInput":
-        return self._sandbox
-
-    @sandbox.setter
-    def sandbox(self, value: "SandboxInput"):
-        self._sandbox = value
-
     def get_system_prompt(self, ctx: AnyContext) -> str:
         if self._prompt_manager is None:
             return ""
@@ -853,43 +829,3 @@ class LLMChatTask(BuilderMixin, RunnerMixin, BaseTask):  # type: ignore[reportIn
         if rendered_model is not None:
             return rendered_model
         return self._llm_config.model
-
-    @property
-    def history_manager(self) -> AnyHistoryManager | None:
-        """Get the history manager."""
-        return self._history_manager
-
-    @history_manager.setter
-    def history_manager(self, value: AnyHistoryManager | None):
-        """Set the history manager."""
-        self._history_manager = value
-
-    @property
-    def ui_factories(self) -> list[Callable[..., "UIProtocol"]]:
-        """Get the UI factories."""
-        return self._ui_factories
-
-    @ui_factories.setter
-    def ui_factories(self, value: list[Callable[..., "UIProtocol"]]):
-        """Set the UI factories."""
-        self._ui_factories = value
-
-    @property
-    def approval_channels(self) -> list["ApprovalChannel"]:
-        """Get the approval channels."""
-        return self._approval_channels
-
-    @approval_channels.setter
-    def approval_channels(self, value: list["ApprovalChannel"]):
-        """Set the approval channels."""
-        self._approval_channels = value
-
-    @property
-    def include_default_ui(self) -> bool:
-        """Check if the default UI should be included."""
-        return self._include_default_ui
-
-    @include_default_ui.setter
-    def include_default_ui(self, value: bool):
-        """Set if the default UI should be included."""
-        self._include_default_ui = value
